devices: MIMX9352: iomuxc: convert to arm64 compatible
Fix some compatible issues on Cortex-A Core.

+++ devices/MIMX9352/drivers/fsl_iomuxc.h
@@ -699,12 +699,12 @@ static inline void IOMUXC_SetPinMux(uint32_t muxRegister,
 {
     if (muxRegister != 0U)
     {
-        *((volatile uint32_t *)muxRegister) = IOMUXC_PAD_MUX_MODE(muxMode) | IOMUXC_PAD_SION(inputOnfield);
+        *((volatile uint32_t *)(uintptr_t)muxRegister) = IOMUXC_PAD_MUX_MODE(muxMode) | IOMUXC_PAD_SION(inputOnfield);
     }
 
     if (inputRegister != 0U)
     {
-        *((volatile uint32_t *)inputRegister) = inputDaisy;
+        *((volatile uint32_t *)(uintptr_t)inputRegister) = inputDaisy;
     }
 }
 /*!
@@ -727,7 +727,7 @@ static inline void IOMUXC_SetPinConfig(uint32_t muxRegister,
 {
     if (configRegister != 0U)
     {
-        *((volatile uint32_t *)configRegister) = configValue;
+        *((volatile uint32_t *)(uintptr_t)configRegister) = configValue;
     }
 }
 /*@}*/
